Refactor main app to Flask
