+++ flask_app/models.py
@@ -1,0 +1,22 @@
+import torch
+from chatterbox.tts import ChatterboxTTS
+from chatterbox.vc import ChatterboxVC
+
+DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
+
+tts_model = None
+vc_model = None
+
+
+def get_tts_model() -> ChatterboxTTS:
+    global tts_model
+    if tts_model is None:
+        tts_model = ChatterboxTTS.from_pretrained(DEVICE)
+    return tts_model
+
+
+def get_vc_model() -> ChatterboxVC:
+    global vc_model
+    if vc_model is None:
+        vc_model = ChatterboxVC.from_pretrained(DEVICE)
+    return vc_model
