+++ flask_app/utils.py
@@ -1,0 +1,11 @@
+import random
+import numpy as np
+import torch
+
+
+def set_seed(seed: int) -> None:
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    random.seed(seed)
+    np.random.seed(seed)
