+++ pyproject.toml
@@ -18,7 +18,8 @@ dependencies = [
     "diffusers==0.29.0",
     "resemble-perth==1.0.1",
     "conformer==0.3.2",
-    "safetensors==0.5.3"
+    "safetensors==0.5.3",
+    "soundfile>=0.13"
 ]
 
 [project.urls]
