+++ src/chatterbox/audio_editing.py
@@ -1,5 +1,4 @@
 import numpy as np
-import librosa
 
 
 def splice_audios(audios):
